Simplify min, max, stdev and mean reducers
