+++ datatable/expr/__init__.py
@@ -1,8 +1,24 @@
-#!/usr/bin/env python3
-# © H2O.ai 2018; -*- encoding: utf-8 -*-
-#   This Source Code Form is subject to the terms of the Mozilla Public
-#   License, v. 2.0. If a copy of the MPL was not distributed with this
-#   file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#!/usr/bin/env python
+#-------------------------------------------------------------------------------
+# Copyright 2018 H2O.ai
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+# IN THE SOFTWARE.
 #-------------------------------------------------------------------------------
 
 from .abs_expr import abs
@@ -14,11 +30,8 @@ from .dtproxy import f, g
 from .exp_expr import exp, log, log10
 from .isna_expr import isna
 from .literal_expr import LiteralExpr
-from .mean_expr import MeanReducer, mean
-from .minmax_expr import MinMaxReducer, min, max
-from .reduce_expr import ReduceExpr, sum, count, first
+from .reduce_expr import ReduceExpr, sum, count, first, mean, min, max, sd
 from .relop_expr import RelationalOpExpr
-from .sd_expr import StdevReducer, sd
 from .unary_expr import UnaryOpExpr
 
 __all__ = (
@@ -41,10 +54,7 @@ __all__ = (
     "NewColumnExpr",
     "BaseExpr",
     "LiteralExpr",
-    "MeanReducer",
-    "MinMaxReducer",
     "ReduceExpr",
     "RelationalOpExpr",
-    "StdevReducer",
     "UnaryOpExpr",
 )
