+++ datatable/expr/reduce_expr.py
@@ -1,23 +1,34 @@
-#!/usr/bin/env python3
-# © H2O.ai 2018; -*- encoding: utf-8 -*-
-#   This Source Code Form is subject to the terms of the Mozilla Public
-#   License, v. 2.0. If a copy of the MPL was not distributed with this
-#   file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#!/usr/bin/env python
 #-------------------------------------------------------------------------------
-# mean_expr, sd_expr, minmax_expr can eventually be merged into here too
-
+# Copyright 2018 H2O.ai
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+# IN THE SOFTWARE.
+#-------------------------------------------------------------------------------
 from .base_expr import BaseExpr
 from .consts import reduce_opcodes, ops_rules, baseexpr_opcodes
+from datatable.frame import Frame
 from datatable.lib import core
 from datatable.types import stype
 
 _builtin_sum = sum
-
-def sum(iterable, start=0):
-    if isinstance(iterable, BaseExpr):
-        return ReduceExpr("sum", iterable)
-    else:
-        return _builtin_sum(iterable, start)
+_builtin_min = min
+_builtin_max = max
 
 
 def count(iterable=None):
@@ -37,6 +48,49 @@ def first(iterable):
             return x
 
 
+def mean(expr):
+    return ReduceExpr("mean", expr)
+
+
+def sd(expr):
+    return ReduceExpr("stdev", expr)
+
+
+# noinspection PyShadowingBuiltins
+def sum(iterable, start=0):
+    if isinstance(iterable, BaseExpr):
+        return ReduceExpr("sum", iterable)
+    else:
+        return _builtin_sum(iterable, start)
+
+
+# noinspection PyShadowingBuiltins
+def min(*args, **kwds):
+    if len(args) == 1 and isinstance(args[0], BaseExpr):
+        return ReduceExpr("min", args[0])
+    elif len(args) == 1 and isinstance(args[0], Frame):
+        return args[0].min()
+    else:
+        return _builtin_min(*args, **kwds)
+
+
+# noinspection PyShadowingBuiltins
+def max(*args, **kwds):
+    if len(args) == 1 and isinstance(args[0], BaseExpr):
+        return ReduceExpr("max", args[0])
+    elif len(args) == 1 and isinstance(args[0], Frame):
+        return args[0].max()
+    else:
+        return _builtin_max(*args, **kwds)
+
+
+sum.__doc__ = _builtin_sum.__doc__
+min.__doc__ = _builtin_min.__doc__
+max.__doc__ = _builtin_max.__doc__
+
+
+
+
 class CountExpr(BaseExpr):
 
     def __str__(self):
@@ -51,7 +105,6 @@ class ReduceExpr(BaseExpr):
     __slots__ = ["_op", "_expr"]
 
     def __init__(self, op, expr):
-        super().__init__()
         self._op = op
         self._expr = expr
